requirement.txt
numpy
torch>=1.7
torchvision
matplotlib
pillow
opencv-python
pandas